chore(architecture): move runtime cloud link policy
