+++ crates/trust-runtime/tests/runtime_cloud_architecture.rs
@@ -11,6 +11,7 @@ fn runtime_cloud_core_modules_do_not_import_transport_layers() {
     let sources = [
         "src/runtime_cloud/contracts.rs",
         "src/runtime_cloud/ha.rs",
+        "src/runtime_cloud/link_policy.rs",
         "src/runtime_cloud/profile_policy.rs",
         "src/runtime_cloud/projection.rs",
         "src/runtime_cloud/routing.rs",
